Rename cache helper file
