+++ packages/react/src/runtime/cache.js.flow
@@ -1,5 +1,5 @@
 /**
- * Flowtype definitions for disable-cache
+ * Flowtype definitions for cache
  * Generated by Flowgen from a Typescript Definition
  * Flowgen v1.15.0
  * @flow
